refactor: time table refactor-previousAvailTime을 기존 선택값과 합치는 로직 다른 파일로 추출
