+++ src/pages/participate/Participate_timetable/Timetable.tsx
@@ -13,20 +13,18 @@ import {
   getDay,
   getHours,
   format as formatDate,
-  setMinutes,
-  setSeconds,
-  setMilliseconds,
   addMinutes,
   isAfter,
-  getTime,
-  getMinutes,
 } from "date-fns";
+import { snap30, keyOf } from "@/utils/dateUtil";
+import { useMergePreviousTimes } from "./TimetableHooks/useMergePreviousTime";
+
 import { ko } from "date-fns/locale";
 
-import type{
+import type {
   ParticipateObject,
   UISlot,
-}from "@apis/participate/participateTypes";
+} from "@apis/participate/participateTypes";
 
 interface BusyInterval {
   startAt: string;
@@ -48,51 +46,17 @@ const Timetable = ({
   scheduleData,
   previousAvailTime,
 }) => {
-const snap30 = (d: Date) => {
-    const m = Math.floor(getMinutes(d) / 30) * 30;
-    return setMilliseconds(setSeconds(setMinutes(d, m), 0), 0);
-  };
-  const keyOf = (sISO: string, eISO: string) => {
-    const s = getTime(setMilliseconds(setSeconds(parseISO(sISO), 0), 0));
-    const e = getTime(setMilliseconds(setSeconds(parseISO(eISO), 0), 0));
-    return `${s}|${e}`;
-  };
-
-  useEffect(() => {
-    if (!Array.isArray(previousAvailTime) || previousAvailTime.length === 0)
-      return;
-
-    setSelectedTimes((prev) => {
-      const exists = new Set(prev.map((s) => keyOf(s.startAt, s.endAt)));
-      const merged = [...prev];
-      let changed = false;
-
-      for (const slot of previousAvailTime) {
-        const s = snap30(parseISO(slot.startAt));
-        const e = snap30(parseISO(slot.endAt));
-        const sISO = s.toISOString();
-        const eISO = e.toISOString();
-        const key = keyOf(sISO, eISO);
-        if (!exists.has(key)) {
-          merged.push({ startAt: sISO, endAt: eISO });
-          exists.add(key);
-          changed = true;
-        }
-      }
-
-      return changed ? merged : prev;
-    });
-  }, [previousAvailTime, setSelectedTimes]);
+  useMergePreviousTimes(previousAvailTime, setSelectedTimes);
   console.log(candidateDates);
   const sortedDates: string[] = [...(candidateDates ?? [])].sort();
   console.log(sortedDates);
   console.log(previousAvailTime);
 
   if (sortedDates.length === 0) return <p>표시할 날짜가 없습니다.</p>;
-  
+
   const validDates: Date[] = sortedDates.map((dateStr) => parseISO(dateStr));
 
-   const start = validDates[0];
+  const start = validDates[0];
   const end = validDates[validDates.length - 1];
   const daysSpan = differenceInCalendarDays(end, start) + 1;
 
@@ -102,7 +66,10 @@ const snap30 = (d: Date) => {
   );
 
   const rangeStart = formatDate(startOfDay(validDates[0]), "yyyy-MM-dd");
-  const rangeEndExclusive = formatDate(addDays(endOfDay(validDates.at(-1)!), 1), "yyyy-MM-dd");
+  const rangeEndExclusive = formatDate(
+    addDays(endOfDay(validDates.at(-1)!), 1),
+    "yyyy-MM-dd"
+  );
 
   const handleSelect = (info: any) => {
     let s = snap30(info.start as Date);
@@ -115,7 +82,7 @@ const snap30 = (d: Date) => {
     const k = keyOf(sISO, eISO);
 
     const exists = selectedTimes.some((t) => keyOf(t.startAt, t.endAt) === k);
-    
+
     if (exists) {
       setSelectedTimes((prev) =>
         prev.filter((t) => keyOf(t.startAt, t.endAt) !== k)
@@ -125,7 +92,7 @@ const snap30 = (d: Date) => {
     }
   };
 
-   const busyEvents = (scheduleData ?? []).map((ev) => ({
+  const busyEvents = (scheduleData ?? []).map((ev) => ({
     start: parseISO(ev.startAt), // 'YYYY-MM-DDTHH:mm:ss' → 로컬 Date로 해석
     end: parseISO(ev.endAt),
     display: "background",
